extract menu actions from vm
